Split BlogImage into client/server component
